# Conflicts:
#	WebApp/src/environments/environment.ts

+++ WebApp/src/app/business-modules/catalogue/partner-data/partner.component.ts
@@ -16,7 +16,7 @@ import { NgProgress } from '@ngx-progressbar/core';
 import { AppList } from 'src/app/app.list';
 import { ExportRepo, CatalogueRepo } from '@repositories';
 import { catchError, finalize } from 'rxjs/operators';
-import { ConfirmPopupComponent } from '@common';
+import { ConfirmPopupComponent, Permission403PopupComponent } from '@common';
 import { ToastrService } from 'ngx-toastr';
 
 type PARTNERDATA_TAB = 'allTab' | 'Customer' | 'Agent' | 'Carrier' | 'Consginee' | 'Shipper';
@@ -37,6 +37,8 @@ enum PartnerDataTab {
 export class PartnerComponent extends AppList implements OnInit {
     @ViewChild(PartnerListComponent, { static: false }) allPartnerComponent: PartnerListComponent;
     @ViewChild(ConfirmPopupComponent, { static: false }) confirmDeletePopup: any;
+    @ViewChild(Permission403PopupComponent, { static: false }) info403Popup: Permission403PopupComponent;
+
     pager: PagerSetting = PAGINGSETTING;
     partnerDataSettings: ColumnSetting[] = PARTNERDATACOLUMNSETTING;
     configSearch: any = {
@@ -145,26 +147,62 @@ export class PartnerComponent extends AppList implements OnInit {
         this.partner = event;
         this.confirmDeletePopup.show();
     }
-    showDetail(event) {
-        this.partner = event;
-        this.router.navigate([`/home/catalogue/partner-data/detail/${this.partner.id}`]);
-    }
-    onDelete() {
-        this._catalogueRepo.deletePartner(this.partner.id)
-            .pipe(catchError(this.catchError), finalize(() => this._progressRef.complete()))
-            .subscribe(
-                (res: CommonInterface.IResult) => {
-                    if (res.status) {
-                        this._toastService.success(res.message);
 
-                        this.allPartnerComponent.dataSearch = this.criteria;
-                        this.allPartnerComponent.getPartners();
-                        this.confirmDeletePopup.hide();
+    showDetail(event) {
+        this.partner = event
+        this._catalogueRepo.checkViewDetailPartnerPermission(this.partner.id)
+            .pipe(
+                catchError(this.catchError),
+                finalize(() => this._progressRef.complete())
+            ).subscribe(
+                (res: any) => {
+                    if (res) {
+                        this.router.navigate([`/home/catalogue/partner-data/detail/${this.partner.id}`]);
                     } else {
-                        this._toastService.error(res.message);
+                        this.info403Popup.show();
                     }
-                }
+                },
             );
+    }
+
+    onDelete() {
+        let allowDelete = false;
+        this._catalogueRepo.checkDeletePartnerPermission(this.partner.id)
+            .pipe(
+                catchError(this.catchError),
+                finalize(() => this._progressRef.complete())
+            ).subscribe(
+                (res: any) => {
+                    if (!res) {
+                        this.info403Popup.show();
+                        this.confirmDeletePopup.hide();
+                        allowDelete = false;
+                        return;
+                    } else {
+                        allowDelete = true;
+                        this._progressRef.start();
+                    }
+                },
+            );
+        if (allowDelete) {
+            this.confirmDeletePopup.hide();
+            this._catalogueRepo.deletePartner(this.partner.id)
+                .pipe(catchError(this.catchError), finalize(() => this._progressRef.complete()))
+                .subscribe(
+                    (res: CommonInterface.IResult) => {
+                        if (res.status) {
+                            this._toastService.success(res.message);
+
+                            this.allPartnerComponent.dataSearch = this.criteria;
+                            this.allPartnerComponent.getPartners();
+                            this.confirmDeletePopup.hide();
+                        } else {
+                            this._toastService.error(res.message);
+                        }
+                    }
+                );
+        }
+
     }
 
     addPartner() {
